[TRTLLM-11318][feat] move VisualGen APIs to a separate dir
