+++ .github/CODEOWNERS
@@ -165,6 +165,7 @@ docs/source/performance/perf-benchmarking.md @NVIDIA/trtllm-bench-reviewers
 /tensorrt_llm/executor @NVIDIA/trt-llm-llmapi-devs
 /tensorrt_llm/serve @NVIDIA/trt-llm-llmapi-devs
 /tensorrt_llm/commands @NVIDIA/trt-llm-llmapi-devs
+/tensorrt_llm/visual_gen @NVIDIA/trt-llm-llmapi-devs
 
 ## TensorRT-LLM LLM Disaggregated
 /examples/disaggregated @NVIDIA/trt-llm-disagg-devs @NVIDIA/trt-llm-doc-owners
