+++ tensorrt_llm/llmapi/__init__.py
@@ -23,13 +23,10 @@ from .llm_utils import (BuildConfig, KvCacheRetentionConfig, QuantAlgo,
                         QuantConfig)
 from .mm_encoder import MultimodalEncoder
 from .mpi_session import MpiCommSession
-from .visual_gen import VisualGen, VisualGenParams
 
 __all__ = [
     'LLM',
     'AsyncLLM',
-    'VisualGen',
-    'VisualGenParams',
     'MultimodalEncoder',
     'CompletionOutput',
     'RequestOutput',
